+++ HallRentingService.WebAPI/Features/Booking/Handlers/BookHall/BookHallResponse.cs
@@ -1,3 +1,3 @@
 ﻿namespace HallRentingService.WebAPI.Features.Booking.Handlers.BookHall;
 
-public sealed record class BookHallResponse(Guid Id, float TotalPrice);
+public sealed record class BookHallResponse(Guid Id, decimal TotalPrice);
