+++ HallRentingService.WebAPI/Features/Halls/Dtos/HallDto.cs
@@ -7,7 +7,7 @@ public sealed class HallDto
     public Guid Id { get; set; }
     public required string Name { get; set; }
     public int Capacity { get; set; }
-    public float PricePerHour { get; set; }
+    public decimal PricePerHour { get; set; }
     public List<HallService_JoinDto> HallServices { get; set; } = [];
     public List<Guid> Bookings { get; set; } = [];
 }
